requirements.txt
flask
python-dotenv
google-generativeai
fastapi
PyJWT
jinja2
python-multipart
google
uvicorn
itsdangerous
motor
typing
pydantic
pydantic[email]